+++ CHANGELOG.md
@@ -13,6 +13,7 @@
     * Fixed broken `environment://` stream when it doesn't have configuration
     * Fixed `Flex Object` missing key field value when using `FolderStorage`
     * Fixed broken Twig try tag when catch has not been defined or is empty
+    * Fixed `FlexForm` serialization
 
 # v1.7.18
 ## 07/19/2021